# packages/shared-types/openapi.json
{
  "components": {
    "schemas": {
      "Catalyst": {
        "additionalProperties": false,
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "kind": {
            "enum": [
              "earnings",
              "macro",
              "corporate",
              "other"
            ],
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "date",
          "description",
          "kind"
        ],
        "title": "Catalyst",
        "type": "object"
      },
      "Citation": {
        "additionalProperties": false,
        "properties": {
          "fetched_at": {
            "format": "date-time",
            "title": "Fetched At",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "title",
          "source",
          "fetched_at"
        ],
        "title": "Citation",
        "type": "object"
      },
      "Entry": {
        "additionalProperties": false,
        "properties": {
          "conditions": {
            "title": "Conditions",
            "type": "string"
          },
          "kind": {
            "enum": [
              "limit",
              "market",
              "stop_limit"
            ],
            "title": "Kind",
            "type": "string"
          },
          "levels": {
            "items": {
              "type": "string"
            },
            "title": "Levels",
            "type": "array"
          }
        },
        "required": [
          "kind",
          "levels",
          "conditions"
        ],
        "title": "Entry",
        "type": "object"
      },
      "ExitLevel": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "enum": [
              "scale_out",
              "time_stop",
              "invalidation"
            ],
            "title": "Kind",
            "type": "string"
          },
          "portion": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portion"
          },
          "price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "trigger"
        ],
        "title": "ExitLevel",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "NoteCreate": {
        "properties": {
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "NoteCreate",
        "type": "object"
      },
      "NoteRead": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "plan_id": {
            "format": "uuid",
            "title": "Plan Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "plan_id",
          "body",
          "created_at",
          "updated_at"
        ],
        "title": "NoteRead",
        "type": "object"
      },
      "Plan": {
        "additionalProperties": false,
        "properties": {
          "capital": {
            "title": "Capital",
            "type": "string"
          },
          "catalysts": {
            "items": {
              "$ref": "#/components/schemas/Catalyst"
            },
            "title": "Catalysts",
            "type": "array"
          },
          "conviction": {
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Conviction",
            "type": "string"
          },
          "entry": {
            "$ref": "#/components/schemas/Entry"
          },
          "exits": {
            "items": {
              "$ref": "#/components/schemas/ExitLevel"
            },
            "title": "Exits",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "horizon": {
            "enum": [
              "intraday",
              "swing",
              "long_term"
            ],
            "title": "Horizon",
            "type": "string"
          },
          "review_cadence": {
            "title": "Review Cadence",
            "type": "string"
          },
          "risk_flags": {
            "items": {
              "$ref": "#/components/schemas/RiskFlag"
            },
            "title": "Risk Flags",
            "type": "array"
          },
          "sizing": {
            "$ref": "#/components/schemas/Sizing"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Sources",
            "type": "array"
          },
          "stop": {
            "$ref": "#/components/schemas/Stop"
          },
          "thesis": {
            "title": "Thesis",
            "type": "string"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "horizon",
          "capital",
          "generated_at",
          "thesis",
          "conviction",
          "entry",
          "sizing",
          "stop",
          "exits",
          "catalysts",
          "risk_flags",
          "review_cadence",
          "sources"
        ],
        "title": "Plan",
        "type": "object"
      },
      "PlanCreateRequest": {
        "properties": {
          "capital": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Capital"
          },
          "horizon": {
            "enum": [
              "intraday",
              "swing",
              "long_term"
            ],
            "title": "Horizon",
            "type": "string"
          },
          "ticker": {
            "maxLength": 16,
            "minLength": 1,
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "horizon",
          "capital"
        ],
        "title": "PlanCreateRequest",
        "type": "object"
      },
      "PlanRevisionRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "diff_json": {
            "additionalProperties": true,
            "title": "Diff Json",
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "plan_id": {
            "format": "uuid",
            "title": "Plan Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "plan_id",
          "created_at",
          "payload",
          "diff_json"
        ],
        "title": "PlanRevisionRead",
        "type": "object"
      },
      "RiskFlag": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "severity": {
            "enum": [
              "info",
              "warn"
            ],
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "severity",
          "code",
          "message"
        ],
        "title": "RiskFlag",
        "type": "object"
      },
      "Sizing": {
        "additionalProperties": false,
        "properties": {
          "R_value": {
            "title": "R Value",
            "type": "string"
          },
          "dollar_exposure": {
            "title": "Dollar Exposure",
            "type": "string"
          },
          "risk_pct": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Pct",
            "type": "number"
          },
          "shares": {
            "minimum": 0.0,
            "title": "Shares",
            "type": "integer"
          }
        },
        "required": [
          "risk_pct",
          "shares",
          "dollar_exposure",
          "R_value"
        ],
        "title": "Sizing",
        "type": "object"
      },
      "Stop": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "enum": [
              "technical",
              "atr",
              "fixed_pct"
            ],
            "title": "Kind",
            "type": "string"
          },
          "price": {
            "title": "Price",
            "type": "string"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          }
        },
        "required": [
          "price",
          "kind",
          "rationale"
        ],
        "title": "Stop",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WatchlistItemCreate": {
        "properties": {
          "ticker": {
            "maxLength": 16,
            "minLength": 1,
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker"
        ],
        "title": "WatchlistItemCreate",
        "type": "object"
      },
      "WatchlistItemRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_plan_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Plan Id"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "ticker",
          "last_plan_id",
          "created_at",
          "updated_at"
        ],
        "title": "WatchlistItemRead",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "StockIt API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/plans": {
      "get": {
        "operationId": "list_plans_plans_get",
        "parameters": [
          {
            "in": "query",
            "name": "ticker",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ticker"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Plan"
                  },
                  "title": "Response List Plans Plans Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Plans",
        "tags": [
          "plans"
        ]
      },
      "post": {
        "operationId": "create_plan_plans_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Plan",
        "tags": [
          "plans"
        ]
      }
    },
    "/plans/{plan_id}": {
      "get": {
        "operationId": "get_plan_plans__plan_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Plan",
        "tags": [
          "plans"
        ]
      }
    },
    "/plans/{plan_id}/notes": {
      "get": {
        "operationId": "list_notes_plans__plan_id__notes_get",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NoteRead"
                  },
                  "title": "Response List Notes Plans  Plan Id  Notes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Notes",
        "tags": [
          "notes"
        ]
      },
      "post": {
        "operationId": "create_note_plans__plan_id__notes_post",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Note",
        "tags": [
          "notes"
        ]
      }
    },
    "/watchlist": {
      "get": {
        "operationId": "list_watchlist_watchlist_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WatchlistItemRead"
                  },
                  "title": "Response List Watchlist Watchlist Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Watchlist",
        "tags": [
          "watchlist"
        ]
      },
      "post": {
        "operationId": "create_watchlist_item_watchlist_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistItemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistItemRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Watchlist Item",
        "tags": [
          "watchlist"
        ]
      }
    },
    "/watchlist/{item_id}": {
      "delete": {
        "operationId": "delete_watchlist_item_watchlist__item_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Watchlist Item",
        "tags": [
          "watchlist"
        ]
      }
    },
    "/watchlist/{item_id}/refresh": {
      "post": {
        "operationId": "refresh_watchlist_item_watchlist__item_id__refresh_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanRevisionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Watchlist Item",
        "tags": [
          "watchlist"
        ]
      }
    }
  }
}